Move String and UITextField extensions to separate files
